extensions: Port chrome.instanceID API to desktop Android
This API is part of Google Cloud Messaging, and is preferred over the
chrome.gcm API. Port the API tests as well.

One JS subtest has to be skipped due to a platform difference. On
Android the GCM implementation is provided by Google Play Services,
not Chrome code. The Android API always returns success when deleting
a token, even if the token doesn't exist.

Bug: 424845641
Change-Id: Ic10db5928136912b080b6fdc722a4bb3cc6d8cd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6643011
Reviewed-by: Michael Wojcicka <mwoj@google.com>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1473836}