Clean up app menu item click handling.
The previous path when from the AppMenuHandlerImpl through the AppMenu
and back to the AppMenuHandlerImpl. This removes the logic from AppMenu
so it does not need to understand how to handle clicks of items in the
list.

This is part of a larger objective to remove the PropertyModel/ModelList
awareness in AppMenu itself as that breaks the MVC patterns.

Bug: None
Change-Id: I01fecbf743721b435b85ecfde61c10fe7a954c83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6641815
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1473862}