[Bookmarks Bar] Move visibility observers to one class for Bookmark Bar
This CL continues the refactor work for Bookmarks Bar visibility and
settings.

With this CL we remove the BookmarkBarSettingProvider, which is nested
in the visibility provider, and allow the visibility provider to keep
track of all things that may in some way affect visibility and notify
observers as needed. This removes an extra layer and the need to pass
around callbacks and extra suppliers. The settings flow and any flow
interested in the overall visibility of the Bookmark Bar are now fully
decoupled and each side does not need to make separate instances of
the setting provider and can instead rely directly on the visibility
provider.

Bug: 424456738
Change-Id: I664a17a958af37556228f966907f0cd6899cfcab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6643354
Reviewed-by: Brandon Wylie <wylieb@google.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1473871}