[PixAL] Open Wallet account linking page
When user accepts the Pix account linking prompt, open the account
linking flow in Wallet app.

The user would be taken to the KYC page or the account linking page
based on whether or not it is the first time they're linking their
account.

Recording: https://screencast.googleplex.com/cast/NDgwMjQ2NjU0NDYxNTQyNHxkOGMzYTY5Ny02Nw
Note: The bottom sheet should be closed when the user takes action in
the prompt. It will be fixed in a future CL.

Bug: 419728706
Change-Id: If7a7733e93e0b2b9e59da2d3d3bf52a99ff57ad1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6639183
Reviewed-by: Siddharth Shah <siashah@chromium.org>
Reviewed-by: Kavita Soni <kavitasoni@chromium.org>
Commit-Queue: Vishwas Uppoor <vishwasuppoor@google.com>
Cr-Commit-Position: refs/heads/main@{#1474068}