[InputVizard] Destroy and recreate InputReceiver for Android 16+.
Due to an Android platform bug which caused crash on calling
AInputReceiver_relesae, a singe AInputReceiver was created and being
used. This results in not transferring sequences once the browser side
InputTransfer token had changed. With the bug having been fixed in
Android 16, with this change now we are destroying and recreating
InputReceiver every time a root compositor frame sink is destroyed. This
would help in improving InputVizard coverage on Android 16+ devices.

Bug: 424169457, 392885482
Change-Id: Ic083f23c4924c9787dff12699f6baf8f32ec8c56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6641727
Commit-Queue: Kartar Singh <kartarsingh@google.com>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474196}