Reland "[Signin][Android] Have AppRestrictionSupplier implement OneshotSupplier"
This reverts commit ebdb4f9d8017fb16152eb035134c9005e7c39a7e.

The previous CL got reverted because it made the test
testSigninFirstRunLoadPointHistograms flaky. This was because the check
for the histogram MobileFre.FromLaunch.PoliciesLoaded would often run
before the policies were retrieved. This has now been fixed by using
histograms.pollInstrumentationThreadUntilSatisfied() that waits for the
histogram to be recorded.

Bug: 385693639
Change-Id: I2f37459d2298d9b289f649f6c28a017fa4f9ff71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6624219
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Lucia Giorgi <quacksort@google.com>
Reviewed-by: Liza Bipin <mlbipin@google.com>
Cr-Commit-Position: refs/heads/main@{#1474239}