[//gpu] Make CreateGpuMemoryBufferImplFromHandle() private with friend
This method should only be called from within ClientSharedImage. This
CL makes it private with ClientSharedImage as a friend.

There are a few violations of this invariant, which we are working to
remove. In one case it seems tricky to friend the templated class that
makes the violating call, so this CL also adds a public wrapper for
exactly that use case with instructions not to add any more usages.

Bug: 404905709
Change-Id: Id4f0884532c625de5a929e86afaadae4797cd524
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6641803
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474262}