[RSG] Move Autofill.Offer.Suggestion2 metric recording to BAM
This CL mimics the crrev.com/c/6551775 but addresses the MPC
suggestion metrics instead.

Before this CL the likelihood that generated MPC suggestions will be
shown (given that they were generated) was high, so we've recorded them as showed before the UI was actually rendered.

As part of the suggestion generation refactoring
(see go/refactor-suggestion-generation) this assumption won't be true
anymore (since multiple FillingProducts will generate suggestions and
the decision which to show will be done later).

This CL moves the suggestion-shown signaling logic to the BAM signal that indicates the popup was shown. This provides greater confidence that MPC suggestions were actually visible.

Bug: 409962888
Change-Id: I24defb79e5dabc5e2383d8e6be8685814f296a2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6641720
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Karol Sygiet <sygiet@google.com>
Reviewed-by: Olivia Saul <jsaul@google.com>
Cr-Commit-Position: refs/heads/main@{#1474287}