Migrate URL activitySource to use NSExtensionItem.
Attributed Content Text was added in crrev.com/c/6439511 as Gmail
requested that we provide a title with the URL. But this change seems to
confuse other share target, like the OS Reminders app, or app extension
that use simplified NSExtensionActivationRule. The other way to provide
a title without sharing multiple items is to share an NSExtensionItem.
This CL migrate the sharing of URLs to NSExtensionItem. That way,
we can provide the title to the share extension.

Sadly, some OS app (like message) do not support it, so special case
them.

Bug: 422707948
Change-Id: I501d96e0c73818513636ddccb0e93fa39513ce82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6633417
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474325}