[media] Fix alpha type for delegation and YUV-RGB conversion for VRU
Currently, performing WritePixels on raster interface avoids going
through direct upload path since color space is not passed in [1], and
instead creates an SkSurface and writes pixels to it (unless the
video color space is also invalid). This should be fixed so that
direct texture upload is preferred.

On handling that for mac, we found that direct texture upload can be
broken for delegated compositing as overlays require premultiplied
alpha as it doesn't go over SkiaRenderer which does the
premultiplication typically. This change handles that as well for
when scanout usage is set and also passes correct alpha bool for
converting YUV video frame to RGB.

Eventually we should enforce in CaLayerOverlay (and maybe even
DcLayerOverlay) that the alpha type is Premul for TextureDrawQuads or
it cannot go to overlay.

[1] https://source.chromium.org/chromium/chromium/src/+/main:gpu/command_buffer/service/raster_decoder.cc;drc=5e74ac5d42bce00b534d992272e7fa09b53b300f;l=2066

Bug: 355923583
Change-Id: I95e87aa95c5750aef94b05567d227b9fbdea1944
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6638968
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474382}