[A11yPerformance] Use isAccessibilityElement() to simplify Mac tree
There are cases to be careful of, such as if something becomes ignored in the parent chain, but in all cross-platform cases of ignored, this
is handled already via PlatformGet() methods skipping the nodes.
In the case of mac-specific ignored cases, the entire subtree will be
ignored (e.g. invisible subtrees, or <img alt=""> which is a leaf).

After this change we no longer fire Mac events for selected children
changed on collapsed combobox popups (AXMenu), because those objects
are invisible. However, there is still a value change event in that
case. This seems sensible because the only child that's exposed in
that case is the currently selected option. There isn't a set of
available children to select.

Bug: none
Change-Id: I1eb198ff96a5fe07792727f2cc4964634f336394
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6639815
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474391}