App-Bound: Explicitly clean up bad or invalid encrypted keys
This CL adds a specific check for 'short key' when retrieving the
encrypted key from storage.

The encrypted key must be at least as large as the expected decrypted
key.

When an encrypted key is invalid due to failing one of the constraints,
this is an irrecoverable error so the key is now deleted from backing
store.

As a drive-by, this CL also corrects a value in an enum that was
previously incorrect.

BUG=423663201

Change-Id: I11a505b24f0733b218b001ef14ed6d5282875e40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6635099
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474443}