Track picture-in-picture occlusion for the external protocol dialog
Currently we prevent input on permission prompts that are occluded by a
picture-in-picture window. The external protocol dialog is not a
permission prompt, and therefore does not get that protection.

This change updates the external protocol dialog to prevent user input
when occluded by a picture-in-picture window. We also maintain input
protection for a short period after the picture-in-picture window is
closed, through the use of the input protector.

This change also updates the DialogDelegate to allow clients to
configure whether to allow or block key events during input protection.

Bug: 414640408, 421404692
Change-Id: Ic12b94db5b4bfd607797c9ed958102511621c284
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6638015
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Benjamin Keen <bkeen@google.com>
Cr-Commit-Position: refs/heads/main@{#1474480}