Correctly re-show hidden nearby sharesheet bubble.
1) The sharesheet bubble may be hidden when the parent window is
minizmied.
2) ARC app may recreate nearby share session when state
changes, which will lead to 're-showing' the hidden shareseet
bubble.
3) SharesheetBubbleViewDelegate::ShowNearbyShareBubbleForArc
assumes that the widget is either visible, otherwise not created, and
it doesn't handle the case that widget is hidden.

This CL fixes it by
  a) if it's created, and hidden reshow it.
  b) if it's already created and visible, same as before.
  c) if it's not created yet, same as before.

The bubble will actually be made visible by TransientWindowManager,
so "Show" isn't necessary in this scenario, but I added it for the
correctness.

It's unclear to me if the similar scenario can happen to web apps ATM.
I'll investigate it separately.

Bug: 418924771
Test: covered by unittest. confirmed that this crashes w/o fix.
Change-Id: I07012b848551245dfc77067f60b76b24ad28500c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6635456
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474520}