[Gap Decorations]: Add more tests for subgrid
This CL adds more tests for gap decorations in subgrid. The changes
include tests with a subgrid occupying a portion of the parent grid,
subgridded on either or both axes. It also includes a test for different
decorations in the parent grid and subgrid. Finally, a test is added for
spanning items existing in both the parent grid and subgrid
using`*rule-break` default value: spanning-item`.

A future CL will include tests related to writing modes added.

Bug: 357648037
Change-Id: I991c995b40dcf05c996820cc626c43e8d05520ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6642524
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1474552}