[owners] Remove eugenex@google.com from ash/webui/metrics/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/eugenex@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ica0ae3cf08aa91e54c1e610c935b08c986f203c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6554526
Commit-Queue: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1474566}