[LNA] Add default IP address space override for browser tests
Add a default IP address space override for all InProcessBrowserTest and
ContentBrowserTest tests for all ports of 127.0.0.1 to be in the public
address space. This is to preserve the existing behaviour for all
current tests that use EmbeddedTestServers to simulate public content
and are not testing IP address space functionality. A few examples of
tests that break:

* Tests that use data urls which fetch subresources from EmbeddedTestServers - example: https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/renderer_context_menu/render_view_context_menu_browsertest.cc;l=351-354;drc=80d60dc23376dece2d1a0bad6ff75e5a8a6aadcc

* Tests using isolated-app:// urls that iframe resources from EmbeddedTestServers - example https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/usb/chrome_usb_browsertest.cc;l=997-998;drc=5fefbd27bb934f4da4d57a3d476f116d45855916

* Extension service workers that fetch data from EmbeddedTestServers - example -https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/extensions/extension_cookies_browsertest.cc;l=803-804;drc=1ce527c58df7c82d80b51d0b646b16f89e8bed59

In all 3 examples, the origins of these requests is all considered to be
origins in the public address space (data URLs, isolated_app://,
chrome-extension://).

To do this, we are changing the behaviour of the
ip-address-space-overrides command line to treat port 0 as "all ports
for that IP address".

This will avoid changes such as https://crrev.com/c/6633493.

There are a few tests already written that do care about the IP address
space of 127.0.0.1, and for those we can overwrite the defaults set for
InProcessBrowserTest and ContentBrowserTest.

This change is in preparation for changing the default of
LocalNetworkAccessChecks in
testing/variations/fieldtrial_testing_config.json to blocking, which
currently breaks many tests.

Bug: 395895368, 418737577
Change-Id: I3c0338bea6060070d37572a3254ec634b4919511
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6638311
Commit-Queue: Hubert Chao <hchao@chromium.org>
Reviewed-by: mmenke <mmenke@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474567}