[infra] Enforce builder_config_settings is not needlessly set.
crrev.com/c/5546319 enabled a module-level default for
builder_config_settings. Prior to that change, it was enforced that
builder_config_settings wasn't set unless builder_spec or mirrors was
set for the builder The check was removed in that change as otherwise
builders that were not using builder configs would cause an error if the
module-level default for builder_config_settings was set.

This changes it so that if the builder does not set builder_spec or
mirrors, then builder_config_settings will be set to None so that only
builders that explicitly set builder_config_settings can produce errors.

Change-Id: Id89114e0fab5ef20db174c3c19270a81b605e24c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6649624
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1474593}