Fix tool use when seeing kExecutorDestroyed
This is happening when a tool starts a navigation and indicates the
RenderFrame in the renderer was destroyed before the browser sees the
RenderFrameHostChanged signal as it should via [1].

It's not clear to me how this happens and I can't reproduce locally -
frame swapping is notoriously complex with lots of variations
(RenderDocument, speculative frames, etc.).

For now it's probably safer to just assume the lost frame is due to
navigation and proceed successfully.

[1] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/actor/tools/page_tool.cc;l=319;drc=903b7e373540ce3673c21b94e8903d962f31d658

Change-Id: Ib6b7e9083f5e1460267f4fcbef8932e1920697ef
Fixed: 423932492
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6647600
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474610}