policy: Clean up EnterpriseHardwarePlatformAPIEnabled for Android
This policy enables an extension API. That API was recently ported to
desktop Android:
https://chromium-review.googlesource.com/c/chromium/src/+/6642319

Remove future-on fuchsia, since fuchsia does not support extensions.
Update the prefs test to test on desktop_android. There's no need to
test on stock android, as it does not support extensions.

In the yaml file the policy is already enabled on "android", which is
what we're using for now for desktop Android (see for example
ExtensionSettings.yaml).

Bug: 405218977
Change-Id: Iaac23886163c080b196649331006977126834385
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6639026
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474638}