Add details to PaymentCredentialInstrument. (2/3)
Before this change the details field on PaymentCredentialInstrument was
passed to SecurePaymentConfirmationAppFactory in
SecurePaymentConfirmationRequest; however,
SecurePaymentConfirmationAppFactory did not use this value. After this
change the SecurePaymentConfirmationAppFactory forwards the details
string to SecurePaymentConfirmationApp where it is available via the
GetSubLabel() method.

The SecurePaymentConfirmation UX changes are behind flag
chrome://flags#enable-secure-payment-confirmation-ux-refresh

Bug: 405173922, 423841925
Change-Id: Iaec13059cff38994d9cb958b16120b70de6eb532
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6637915
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Slobodan Pejic <slobodan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474643}